requirements.txt
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
Pillow==10.1.0
spidev==3.6
RPi.GPIO==0.7.1
